requirements.txt
PySimpleGUI==4.19.0
dill==0.3.1.1
gspread==3.6.0
oauth2client==4.1.3
python-dateutil==2.8.1